backend/GetResumeCounter/__init__.py
import azure.functions as func

def main(req: func.HttpRequest) -> func.HttpResponse:
    # TODO: Connect to Cosmos DB and increment visitor count
    return func.HttpResponse(
        "Visitor counter function placeholder",
        status_code=200
    )
